+++ minus_games_gui/src/minus_games_gui/views/ready.rs
@@ -4,5 +4,5 @@ use crate::minus_games_gui::style_constants::MARGIN_DEFAULT;
 use iced::widget::{Row, row};
 
 pub(crate) fn view(minus_games_gui: &MinusGamesGui) -> Row<MinusGamesGuiMessage> {
-    row![minus_games_gui.create_ready_view()].padding(MARGIN_DEFAULT)
+    row![minus_games_gui.create_ready_view()].padding(MARGIN_DEFAULT as u16)
 }
